backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
typing-extensions==4.8.0
requests==2.31.0
sqlalchemy==2.0.22
sqlalchemy-utils==0.41.1
passlib==1.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
bcrypt==4.0.1
alembic==1.12.1
httpx==0.25.0
azure-storage-blob==12.18.3
azure-identity==1.14.1
pandas==2.1.1
numpy==1.26.0
python-dateutil==2.8.2
pytz==2023.3
ujson==5.8.0
email-validator==2.1.0
pyyaml==6.0.1
pyotp==2.9.0
qrcode==7.4.2
pillow==10.1.0
cryptography==41.0.4
psutil==5.9.5